+++ src/inspect_scout/_util/path_str.py
@@ -1,17 +1,24 @@
-"""Branded string type for filesystem paths."""
+"""PathStr conversion functions."""
 
 from pathlib import Path
-from typing import NewType
+from urllib.parse import unquote
 
 from upath import UPath
 
-PathStr = NewType("PathStr", str)
-"""A filesystem path as a string (no encoding)."""
+from .str_types import PathStr, UriStr
+
+__all__ = ["PathStr", "as_path", "make_path"]
 
 
-def make_path(path: Path | UPath) -> PathStr:
-    """Convert a Path or UPath to PathStr."""
-    return PathStr(str(path))
+def make_path(path: Path | UPath | UriStr) -> PathStr:
+    """Convert a Path, UPath, or UriStr to PathStr.
+
+    If passed a UriStr, decodes percent-encoded characters.
+    """
+    if isinstance(path, (Path, UPath)):
+        return PathStr(str(path))
+    # UriStr case - decode percent-encoding
+    return PathStr(unquote(UPath(path).path))
 
 
 def as_path(raw: str) -> PathStr:
