+++ src/inspect_scout/_util/uri_str.py
@@ -1,15 +1,12 @@
-"""Branded string type for URIs (percent-encoded per RFC 3986)."""
+"""UriStr conversion functions."""
 
 from pathlib import Path
-from typing import NewType
-from urllib.parse import unquote
 
 from upath import UPath
 
-from .path_str import PathStr, as_path
+from .str_types import PathStr, UriStr
 
-UriStr = NewType("UriStr", str)
-"""A percent-encoded URI string per RFC 3986."""
+__all__ = ["UriStr", "as_uri", "make_uri"]
 
 
 def make_uri(path: Path | UPath | PathStr) -> UriStr:
@@ -18,11 +15,6 @@ def make_uri(path: Path | UPath | PathStr) -> UriStr:
     Resolves the path to an absolute path before encoding.
     """
     return UriStr(UPath(path).resolve().as_uri())
-
-
-def uri_to_path(uri: UriStr) -> PathStr:
-    """Convert a file:// URI to a decoded path string."""
-    return as_path(unquote(UPath(uri).path))
 
 
 def as_uri(raw: str) -> UriStr:
